requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
django-environ==0.9.0

# Shell
ipython==8.12.0
django-extensions==3.2.1

# Redis
django-redis==5.2.0                         # https://pypi.org/project/django-redis/
redis==4.5.3                                # https://pypi.org/project/redis/

# Elastic Search
django-elasticsearch-dsl==7.3               # https://pypi.org/project/django-elasticsearch-dsl/
# elasticsearch==7.13.4
# elasticsearch-dsl==7.4.0

# Pytest
pytest==7.2.2                               # https://github.com/pytest-dev/pytest
pytest-django==4.5.2                        # https://github.com/pytest-dev/pytest-django
pytest-xdist==3.2.1                         # https://github.com/pytest-dev/pytest-xdist
pytest-factoryboy==2.5.1                    # https://github.com/pytest-dev/pytest-factoryboy
pytest-sugar==0.9.6                         # https://github.com/Teemu/pytest-sugar/
pytest-ordering==0.6                        # https://github.com/ftobia/pytest-ordering
tox==4.4.11                                 # https://github.com/tox-dev/tox

# Faker for Fake Data
Faker==18.4.0

# Python Type hint
mypy==1.2.0                                 # https://github.com/python/mypy